// backend/src/db/db.types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Cards {
  id: Generated<number>;
  round_id: number;
  team_id: number;
  word: string;
}

export interface Clues {
  clue: string;
  clue_number: number;
  created_at: Generated<Timestamp | null>;
  id: Generated<number>;
  player_id: number;
  turn_id: number;
}

export interface Games {
  created_at: Generated<Timestamp | null>;
  id: Generated<number>;
  public_id: string | null;
  status_id: number;
}

export interface GameStatus {
  id: Generated<number>;
  status_name: string;
}

export interface Guesses {
  card_id: number;
  created_at: Generated<Timestamp | null>;
  id: Generated<number>;
  player_id: number;
  turn_id: number;
}

export interface PlayerRoles {
  id: Generated<number>;
  role_name: string;
}

export interface PlayerRoundRoles {
  assigned_at: Generated<Timestamp | null>;
  player_id: number;
  role_id: number;
  round_id: number;
}

export interface Players {
  game_id: number;
  id: Generated<number>;
  status_id: number;
  status_last_changed: Generated<Timestamp | null>;
  team_id: number | null;
  user_id: number;
}

export interface PlayerStatuses {
  id: Generated<number>;
  status_name: string;
}

export interface Rounds {
  created_at: Generated<Timestamp | null>;
  game_id: number;
  id: Generated<number>;
  round_number: number;
}

export interface Teams {
  id: Generated<number>;
  team_name: string;
}

export interface Turns {
  created_at: Generated<Timestamp | null>;
  id: Generated<number>;
  round_id: number;
  team_id: number;
}

export interface Users {
  created_at: Generated<Timestamp | null>;
  id: Generated<number>;
  username: string;
}

export interface DB {
  cards: Cards;
  clues: Clues;
  game_status: GameStatus;
  games: Games;
  guesses: Guesses;
  player_roles: PlayerRoles;
  player_round_roles: PlayerRoundRoles;
  player_statuses: PlayerStatuses;
  players: Players;
  rounds: Rounds;
  teams: Teams;
  turns: Turns;
  users: Users;
}